split luci_common into multiple files; rename ToolExit
